index: Extract `repo` module
